+++ lib/data/dynamic_factories.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:vit_gpt_dart_api/data/interfaces/audio_recorder_model.dart';
 import 'package:vit_gpt_dart_api/data/interfaces/completion_model.dart';
 import 'package:vit_gpt_dart_api/data/interfaces/local_storage/local_storage_model.dart';
+import 'package:vit_gpt_dart_api/data/interfaces/realtime_model.dart';
 import 'package:vit_gpt_dart_api/data/interfaces/tts_model.dart';
 
 import 'interfaces/simple_audio_player_model.dart';
@@ -39,6 +40,10 @@ class DynamicFactories {
     String assistantId,
     String threadId,
   )? assistantRepository;
+
+  static RealtimeModel Function()? realtime;
+
+  static CompletionModel Function()? completion;
 }
 
 T _create<T>(T Function()? fac, String name) {