+++ lib/data/interfaces/index.dart
@@ -3,6 +3,7 @@ export 'audio_recorder_model.dart';
 export 'completion_model.dart';
 export 'image_generation_model.dart';
 export 'local_storage/index.dart';
+export 'realtime_model.dart';
 export 'simple_audio_player_model.dart';
 export 'threads_model.dart';
 export 'transcriber_model.dart';