+++ lib/factories/create_realtime_repository.dart
@@ -1,0 +1,11 @@
+import 'package:vit_gpt_dart_api/data/dynamic_factories.dart';
+import 'package:vit_gpt_dart_api/data/interfaces/realtime_model.dart';
+import 'package:vit_gpt_dart_api/repositories/openai/openai_realtime_repository.dart';
+
+RealtimeModel createRealtimeRepository() {
+  var fn = DynamicFactories.realtime;
+  if (fn != null) {
+    return fn();
+  }
+  return OpenaiRealtimeRepository();
+}