+++ lib/factories/index.dart
@@ -1,3 +1,4 @@
 export 'create_image_generation_repository.dart';
+export 'create_realtime_repository.dart';
 export 'create_threads_repository.dart';
 export 'create_transcriber_repository.dart';